requirements.txt
asgiref==3.7.2
asttokens==2.4.1
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
executing==2.0.1
filelock==3.9.0
fonttools==4.45.1
fsspec==2023.4.0
idna==3.4
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.8
networkx==3.0
numpy==1.24.1
packaging==23.2
parso==0.8.3
Pillow==9.3.0
platformdirs==4.0.0
prompt-toolkit==3.0.41
psutil==5.9.6
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
sympy==1.12
torch==2.1.1+cu121
torchaudio==2.1.1+cu121
torchvision==0.16.1+cu121
tornado==6.3.3
traitlets==5.14.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
wcwidth==0.2.12
